changed html layout'

+++ Laravel/laravelwebsite/app/Http/Controllers/PostsController.php
@@ -7,6 +7,11 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -72,7 +77,13 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        
         $post = Post::find($id);
+
+        # Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Post Created');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -108,6 +119,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        # Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Post Created');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
